Text item scroll indicator
Fixed an issue where the scroll indicator was not showing in some scenario. It was caused by the api ref either not being set (on initial render) or holding stale values (when switching between editor/viewer). Instead of computing the showScrollIndicator value in the render the is now a hook that keeps the value in a state and also allows for the value to be recomputed (when the RTE `onCreate` gets called)`